backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
sqlalchemy==2.0.25
aiosqlite==0.19.0
aiomysql==0.2.0
pymysql==1.1.0
pydantic>=2.9.0
pydantic-settings==2.1.0
celery==5.3.6
redis==5.0.1
python-multipart==0.0.6
minio==7.2.3
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
httpx>=0.27.0
cryptography==42.0.0
aiofiles==23.2.1
pandas==2.2.0
